+++ include/rom_8094928.h
@@ -1,7 +1,7 @@
 #ifndef GUARD_ROM_8094928_H
 #define GUARD_ROM_8094928_H
 
-void sub_8094978(u8, int);
+void sub_8094978(u8, u8);
 void sub_8094998(u8[3], u8);
 void sub_8094A74(u8[3], u8, u32);
 u8 sub_8094C20();
